+++ src/Reader/VideoDecoder.cpp
@@ -79,12 +79,12 @@ void VideoDecoder::CheckFlushPacket() {
     if (m_packetQueue.empty()) return;
 
     auto packet = m_packetQueue.front();
-    if (packet->flags & AVFrameFlag::kFlush) {
+    if (packet->flags & static_cast<int>(AVFrameFlag::kFlush)) {
         m_packetQueue.pop_front();
         avcodec_flush_buffers(m_codecContext);
 
         auto videoFrame = std::make_shared<IVideoFrame>();
-        videoFrame->flags |= AVFrameFlag::kFlush;
+        videoFrame->flags |= static_cast<int>(AVFrameFlag::kFlush);
         std::lock_guard<std::recursive_mutex> lock(m_listenerMutex);
         if (m_listener) m_listener->OnNotifyVideoFrame(videoFrame);
     }
@@ -222,7 +222,7 @@ void VideoDecoder::Decode(std::shared_ptr<IAVPacket> packet) {
         return;
     }
     std::lock_guard<std::mutex> lock(m_packetQueueMutex);
-    if (packet->flags & AVFrameFlag::kFlush) {
+    if (packet->flags & static_cast<int>(AVFrameFlag::kFlush)) {
         m_packetQueue.clear();
     }
     m_packetQueue.push_back(packet);
